Remove superfluous debug printout
This printout never inticated error, or bug.
It indicated that given SwitchId was not yet registered in
global ETS table, and subsystems that wanted to communicate with this
switch have to wait until it is fully initialized.
Removed because it was confusing for people not familiar with internal
implementation of LINC.
Fixes #75