Fix add_group test
The time check was incorrect.  We need to make another stats request
after the sleep to get the new elapsed time, instead of retrieving it
from the record we got earlier.

Also made the assertion print the values in question on failure, and
fixed the seconds-to-microseconds conversion (which made no difference
to this test).