Implement reserved OFP port FLOOD
FLOOD port when used as an output port sends the packet out all standard
ports, but not the ingress port, nor ports that are in BLOCKED state
(OpenFlow spec. v1.3.2 ch. 4.5).
This behavior is required as NOX 1.3 controller uses this FLOOD port.
This commit should be reverted as soon as the issue
https://github.com/CPqD/nox13oflib/issues/3 is fixed.
(Motivated by #171).