Add OFPVID_PRESENT bit to VLAN match field in packet-in message
This commit introduces setting the 13rd bit in the oxm_value of match
field for VLAN in packet-in message. For example if the VLAN ID in
a packet has a value of 0x0050 the resulting oxm_value in the match
field will have (0x1000 AND 0x0050). The change affects linc_us{3,4,5}.

Setting the match field for VLAN is optional in packet-in message
and is NOT explicitly defined in the OFP 1.{2,3,4} specifications.

Motivated by #328.