Add reporting OS interfaces status to us4
Physical interfaces are monitored through netlink utility. It notifies
OF ports about `operstate` changes of the underlying physical ports.
The OF port sets its state appropriately and notifies controllers
sending port status message. More about the operstate can be found here:
https://www.kernel.org/doc/Documentation/networking/operstates.txt

OF port has a flag `link_down` that indicates that a physical link is
not present in the port. We assume that the OF port is up (i.e. this flag
is not set) only if the operstate of the port is `up`.